requirements.txt
Flask>=2.3.0
Flask-Babel>=3.1.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
WTForms>=3.0.0
pyotp>=2.8.0
qrcode[pil]>=7.4.0
gunicorn>=21.0.0
gevent>=23.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
cryptography>=41.0.0